vdaf: Move prio3 to its own module
This makes room for other VDAFs we may want to implement.
